PR1 review - updated POMs - replaced mysql by mariadb dependency - removed unnecessaryc comments - updated AuthClient.java - changed Long id to Integer id

+++ material-service/src/main/java/rs/enadzornik/materialservice/client/AuthClient.java
@@ -9,11 +9,5 @@ import rs.enadzornik.materialservice.dto.KorisnikDto;
 @FeignClient(name = "auth-service", url = "http://localhost:8083")
 public interface AuthClient {
     @GetMapping("/api/v1/auth/internal/korisnik/{id}")
-    KorisnikDto getKorisnikById(@PathVariable("id") Long id);
+    KorisnikDto getKorisnikById(@PathVariable("id") Integer id);
 }
-
-//@FeignClient(name = "auth-service", url = "${auth.service.url:http://localhost:8081}")
-//public interface AuthClient {
-//    @GetMapping("/api/v1/auth/internal/korisnik/{id}")
-//    KorisnikDto getKorisnikById(@PathVariable("id") Long id);
-//}
